requirements.txt
torch
scipy
numpy
pandas
xarray
matplotlib
tqdm
PyYAML
rich
transformers
keras